# project/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/user": {
            "post": {
                "description": "create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ReqCreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.responseWithoutData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/:id": {
            "get": {
                "description": "query list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.responseWithoutData"
                        }
                    }
                }
            },
            "put": {
                "description": "update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ReqCreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.responseWithoutData"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.responseWithoutData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ReqLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.responseWithoutData"
                        }
                    }
                }
            }
        },
        "/api/v1/user/refresh": {
            "post": {
                "description": "refresh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.responseWithoutData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.response-array_po_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/po.User"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.response-po_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/po.User"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.responseWithoutData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "po.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "create time",
                    "type": "integer"
                },
                "deleted_at": {
                    "description": "delete time",
                    "type": "integer"
                },
                "email": {
                    "description": "email",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "mobile": {
                    "description": "mobile",
                    "type": "string"
                },
                "nick_name": {
                    "description": "nick_name",
                    "type": "string"
                },
                "password": {
                    "description": "password",
                    "type": "string"
                },
                "updated_at": {
                    "description": "update time",
                    "type": "integer"
                },
                "user_name": {
                    "description": "user_name",
                    "type": "string"
                }
            }
        },
        "vo.ReqCreateUser": {
            "type": "object",
            "required": [
                "email",
                "mobile",
                "nick_name",
                "password",
                "user_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 6
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "vo.ReqLogin": {
            "type": "object",
            "required": [
                "password",
                "user_name"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 6
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    }
}